+++ src/waka-worker/standalone.ts
@@ -1,7 +1,7 @@
 import * as express from 'express'
 import 'dotenv'
 import * as bodyParser from 'body-parser'
-import EnvMapper from '../envMapper'
+import EnvMapper from '../../envMapper'
 import WakaWorker from './index'
 import AWSXRay from 'aws-xray-sdk'
